Refactor Task entity to use Long for ID; update TaskController, TaskRepo, and TaskService accordingly. Add unit tests for TaskRepo and TaskService, and configure H2 database for testing.

+++ src/main/java/org/sebas/taskmanager/controller/TaskController.java
@@ -26,7 +26,7 @@ public class TaskController {
     }
 
     @DeleteMapping("/delete-task/{id}")
-    public void deleteTask(@PathVariable int id){
+    public void deleteTask(@PathVariable Long id){
        service.deleteTask(id);
     }
